Tests running for power of attorney requests controller
