--- modules/representation_management/spec/requests/representation_management/v0/power_of_attorney_submission_spec.rb
@@ -1,188 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-RSpec.describe 'RepresentationManagement::V0::PowerOfAttorneyRequestsController', type: :request do
-  describe 'POST #create' do
-    let(:base_path) { '/representation_management/v0/power_of_attorney_requests' }
-    let(:organization) { create(:organization) } # This is the legacy organization
-    let(:representative) { create(:representative) } # This is the legacy representative
-    let(:params) do
-      {
-        power_of_attorney_request: {
-          record_consent: '',
-          consent_address_change: '',
-          consent_limits: [],
-          claimant: {
-            date_of_birth: '1980-12-31',
-            relationship: 'Spouse',
-            phone: '5555555555',
-            email: 'claimant@example.com',
-            name: {
-              first: 'John',
-              middle: 'Middle', # This is a middle name as submitted by the frontend
-              last: 'Claimant'
-            },
-            address: {
-              address_line1: '123 Fake Claimant St',
-              address_line2: '',
-              city: 'Portland',
-              state_code: 'OR',
-              country: 'USA', # This is a 3 character country code as submitted by the frontend
-              zip_code: '12345',
-              zip_code_suffix: '6789'
-            }
-          },
-          veteran: {
-            ssn: '123456789',
-            va_file_number: '123456789',
-            date_of_birth: '1980-12-31',
-            service_number: '123456789',
-            service_branch: 'ARMY',
-            phone: '5555555555',
-            email: 'veteran@example.com',
-            name: {
-              first: 'John',
-              middle: 'Middle', # This is a middle name as submitted by the frontend
-              last: 'Veteran'
-            },
-            address: {
-              address_line1: '123 Fake Veteran St',
-              address_line2: '',
-              city: 'Portland',
-              state_code: 'OR',
-              country: 'USA', # This is a 3 character country code as submitted by the frontend
-              zip_code: '12345',
-              zip_code_suffix: '6789'
-            }
-          },
-          representative: {
-            organization_id: organization.poa,
-            id: representative.representative_id
-          }
-        }
-      }
-    end
-
-    context 'When submitting all fields with valid data' do
-      before do
-        params[:power_of_attorney_request][:veteran][:service_number] = nil # TEMPORARY FOR FRONTEND TESTING
-        post(base_path, params:)
-      end
-
-      it 'responds with a ok status' do
-        expect(response).to have_http_status(:ok)
-      end
-
-      it 'responds with the expected message' do
-        expect(response.body).to eq({ message: 'Email enqueued' }.to_json)
-      end
-    end
-
-    context 'when submitting with a veteran service number - TEMPORARY FOR FRONTEND TESTING' do
-      before do
-        post(base_path, params: params)
-      end
-
-      it 'responds with the unprocessable_entity status' do
-        expect(response).to have_http_status(:unprocessable_entity)
-      end
-
-      it 'responds with the expected body' do
-        expect(response.body).to eq({ errors: ['render_error_state_for_failed_submission'] }.to_json)
-      end
-    end
-
-    context 'When submitting a valid request without a claimant' do
-      before do
-        params[:power_of_attorney_request][:veteran][:service_number] = nil # TEMPORARY FOR FRONTEND TESTING
-        params[:power_of_attorney_request].delete(:claimant)
-        post(base_path, params:)
-      end
-
-      it 'responds with a ok status' do
-        expect(response).to have_http_status(:ok)
-      end
-
-      it 'responds with a PDF' do
-        expect(response.body).to eq({ message: 'Email enqueued' }.to_json)
-      end
-    end
-
-    context 'When submitting a valid request without a representative id' do
-      before do
-        params[:power_of_attorney_request][:veteran][:service_number] = nil # TEMPORARY FOR FRONTEND TESTING
-        params[:power_of_attorney_request][:representative].delete(:id)
-        post(base_path, params:)
-      end
-
-      it 'responds with a ok status' do
-        expect(response).to have_http_status(:ok)
-      end
-
-      it 'responds with a PDF' do
-        expect(response.body).to eq({ message: 'Email enqueued' }.to_json)
-      end
-    end
-
-    context 'When submitting a valid request with the accredited organization and individual ids' do
-      before do
-        params[:power_of_attorney_request][:veteran][:service_number] = nil # TEMPORARY FOR FRONTEND TESTING
-        accredited_organization = create(:accredited_organization)
-        accredited_individual = create(:accredited_individual)
-        params[:power_of_attorney_request][:representative][:organization_id] = accredited_organization.id
-        params[:power_of_attorney_request][:representative][:id] = accredited_individual.id
-        post(base_path, params:)
-      end
-
-      it 'responds with a ok status' do
-        expect(response).to have_http_status(:ok)
-      end
-
-      it 'responds with a PDF' do
-        expect(response.body).to eq({ message: 'Email enqueued' }.to_json)
-      end
-    end
-
-    context 'when triggering validation errors' do
-      context 'when submitting without the veteran first name for a single validation error' do
-        before do
-          params[:power_of_attorney_request][:veteran][:service_number] = nil # TEMPORARY FOR FRONTEND TESTING
-          params[:power_of_attorney_request][:veteran][:name][:first] = nil
-          post(base_path, params:)
-        end
-
-        it 'responds with an unprocessable entity status' do
-          expect(response).to have_http_status(:unprocessable_entity)
-        end
-
-        it 'responds with the expected body' do
-          expect(response.body).to eq({ errors: ["Veteran first name can't be blank"] }.to_json)
-        end
-      end
-
-      context 'when submitting without multiple required attributes' do
-        before do
-          params[:power_of_attorney_request][:veteran][:service_number] = nil # TEMPORARY FOR FRONTEND TESTING
-          params[:power_of_attorney_request][:veteran][:name][:first] = nil
-          params[:power_of_attorney_request][:veteran][:ssn] = nil
-          params[:power_of_attorney_request][:veteran][:name][:last] = nil
-          params[:power_of_attorney_request][:veteran][:date_of_birth] = nil
-          post(base_path, params:)
-        end
-
-        it 'responds with an unprocessable entity status' do
-          expect(response).to have_http_status(:unprocessable_entity)
-        end
-
-        it 'responds with the expected body' do
-          expect(response.body).to include("Veteran first name can't be blank")
-          expect(response.body).to include("Veteran social security number can't be blank")
-          expect(response.body).to include('Veteran social security number is invalid')
-          expect(response.body).to include("Veteran last name can't be blank")
-          expect(response.body).to include("Veteran date of birth can't be blank")
-        end
-      end
-    end
-  end
-end
